add a second strategy for rendering the diagram - separate window
Previously we only rendered the diagram in the same buf as the plantuml
code. Now we support rendering/updating the diagram in a separate
window.

Add `g:slumlord_separate_win` option to activate this feature.

In future, the "in place" rendering method may be removed to keep things
simple - and because I'm starting to think its just superior.